Add unicode support
Class:Inspector had to remove "Test::Spelling" test completely, because
the test was failing due to a contributor whose name had a none a-zA-Z
character. Adding it as a stopword was not working either.

Pod::Spell documentation @ https://metacpan.org/pod/Pod::Spell#ENCODINGS
notes that default parse_from_file does not apply any encoding layer.
To get the encoding, we should replace it with parse_from_filehandle.

After this change, adding a stopword with non a-zA-Z character works
as expected. I also added a test to make sure this works.